### 2017-07-19 #136 @mstrimas
auk: eBird Data Extraction with AWK
Access to the eBird database, consisting of over 400 million observations, is provided via a huge (>150 GB) text file. The `auk` package extracts records from this file and imports them into R for analysis. Both presence only and presence/absence data can be generated.

```
Package: auk
Title: eBird Data Extraction with AWK
Version: 0.0.2.900
  person("Matthew", "Strimas-Mackey", email = "mes335@cornell.edu", role = c("aut", "cre")),
  person("Eliot", "Miller", role = "aut"),
  person("Wesley", "Hochachka", role = "aut"),
  person("Cornell Lab of Ornithology", role = "cph")
  )
Description: Extract and process bird sightings records from eBird
    (<http://ebird.org>), an online tool for recording bird observations.
    Public access to the full eBird database is via the eBird Basic Dataset
    (EBD; see <http://ebird.org/ebird/data/download> for access), a downloadable
    text file. This package is an interface to AWK for extracting data from the
    EBD based on taxonomic, spatial, or temporal filters, to produce a
    manageable file size that can be imported into R.
    assertthat,
    stringi,
    countrycode,
    tidyr
    covr
```

-   URL for the package (the development repository, not a stylized html page): https://github.com/CornellLabofOrnithology/auk

- Please indicate which category or categories from our [package fit policies](https://github.com/ropensci/onboarding/blob/master/policies.md#package-fit) this package falls under **and why**(? (e.g., data retrieval, reproducibility. If you are unsure, we suggest you make a pre-submission inquiry.):

This package falls somewhere at the intersection of data retrieval and extraction. It provides access to the eBird database; however, it does so by processing a text file downloaded from eBird that contains the full database.

Anyone looking to work with eBird data for science or conservation.

`rebird` provides access to eBird data via the eBird API; however, this only gives access to last 30 days of data. This package is the only one giving access to full eBird database.
