### 2017-07-21 44485a4
add dplyr dependency; remove setclass option; add taxonomic rollup function closes issue #3
