+++ R/auk-package.r
@@ -1,4 +1,4 @@
-#' `auk`: eBird Data Extraction with AWK
+#' `auk`: eBird Data Extraction and Processing with AWK
 #'
 #' Tools for extracting and processing eBird data from the eBird Basic Dataset 
 #' (EBD).
@@ -6,4 +6,5 @@
 #' @name auk
 #' @docType package
 #' @importFrom stringr str_interp
+#' @importFrom rlang .data
 NULL

[…]
