+++ R/auk-rollup.r
@@ -5,7 +5,7 @@
 #' In some cases, a checklist may contain multiple records for the same species, 
 #' for example, both Audubon and Myrtle Yellow-rumped Warblers, as well as some 
 #' records that are not resolvable to species, for example, "warbler sp.". 
-#' This function, addresses these cases by removing taxa not identifiable to 
+#' This function addresses these cases by removing taxa not identifiable to 
 #' species and rolling up taxa identified below species level to a single record 
 #' for each species in each checklist. By default, [read_ebd()] calls 
 #' `ebd_rollup()` when importing an EBD file.
