### 2017-07-25 4472602
Create CODE_OF_CONDUCT.md
