### 2017-07-31 15e27e7
goodpractice::gp() fixes
