+++ DESCRIPTION
@@ -1,7 +1,6 @@
 Package: auk
 Title: eBird Data Extraction and Processing with AWK
-Version: 0.0.2.900
-Date: 2017-07-05
+Version: 0.0.2.901
 Authors@R: c(
   person("Matthew", "Strimas-Mackey", email = "mes335@cornell.edu", role = c("aut", "cre")),
   person("Eliot", "Miller", role = "aut"),
