### 2017-07-31 #136 @mstrimas
Thanks, @karthik, I just fixed several of these, but the following remain:

- Test coverage: not sure where the 80% number comes from, I have [93% coverage on codecov.io](https://codecov.io/github/CornellLabofOrnithology/auk?branch=master), which I believe is good relative to other rOpenSci packages
- For the `R CMD check` warning and error: these don't arise when I use `devtools::check()` or when I submit to CRAN, and I haven't had any luck figuring out what the issue is or how to resolve it
