### 2017-08-24 #136 @emhart
---

The authors present a very elegant solution to a difficult problem in R, how to handle a very large data set such as the eBird data (larger than most people's personal computers could load into RAM) when most users only need a subset of the data in the full file. Their solution is to provide a way to automatically build an Awk script, execute it, and write a new output file. While this could be done without this R package, they make the data accessible to a much larger audience.

Overall I found the code to very comprehensive and elegant and think it will be a good addition to the rOpenSci package suite. The authors largely adhere to the [rOpenSci package guidlines](https://github.com/ropensci/onboarding/blob/master/packaging_guide.md) and are exceedingly diligent in their error handling in each function. Also I was impressed with their coverage use cases in their tests. They went above in beyond in writing an exhaustive suite of tests for each function. I find no major issues with how their code is written.

I do think there are a couple minor areas for improvement in making it easier for end users. The biggest issue I had was initially grokking that this was a multi-step workflow that involved writing a file to disk. My first impression was that I could simlply run a bunch of filters and the `ebd` variable (in the README) would actually be a dataframe. If there was a way to make the work flow more explicit, especially in the README, I think that would be helpful. Another thought I have is, would it be possible to obfuscate this multi-step process and have a function that loads up the ebd, runs the filters, writes the file and reads it all into a tibble?  That way the end user could side-step ever running `read_ebd()`.  Anoher minor issue I had was that there's somewhat mixed handling of what I think of as "user standards laziness". For instance, you insist on ISO date standards, but countries can be mixed case, and don't require the ISO country code. I see how on the one hand you're making it easier for users, but I found myself a bit confused about where I could skip on my standards when it came to input. I was honestly surprised when I could ender "gray jay" but not "Robin" (but "American Robin" worked fine).

*Minor comments*

* Throughout your vignette and README you specify units for everything, but not extent. While I assume that it's decimal degrees, I think it would be good to be explicit.

* You note that for this to work with windows they need cygwin installed in a specific directiory. I could forsee this being difficult, is there any way to specify the path to make it easier for windows users?

* Consider adding a CITATION file so your package can be [cited](http://r-pkgs.had.co.nz/inst.html#inst-citation).

* Would it be too much to include a column filter option to your workflow? Maybe that's a 0.0.3 feature, but seems like it would be a nice addition

* There were some instances where you used system variable names as local variable names in functions, e.g. line 36 in auk_time.r `time <- paste0(ifelse(nchar(time) == 4, "0", ""), time)`. It obviously doesn't cause an issue as is, but maybe it could down the line.

**Community guidelines**

A CONTRIBUTING or way to cotribute in the README is not present. Consider adding contributor guidelines.

**Examples**

I ran all examples using `devtools::run_examples` and all ran without error.

**Tests**

I ran all tests with `devtools::test()` and all tests were passed.

**Checks**

I built the package on the following system using `devtools::test(cran = TRUE)`:
Platform: x86_64-apple-darwin15.6.0 (64-bit)
Running under: macOS Sierra 10.12.6

All checks were passed with no notes, errors, or warninging.

**Test coverage**

I checked for the amount of test coverage using `covr::package_coverage()` and it was 80.9%

Furthermore I reviewed all the tests in `tests/testthat`, not only was there good test coverage, the range of scenarios was exhaustive. I was very impressed with the breadth of cases tested.

**sessionInfo()**

Just so you can see what versions of packages I used to run my tests:

```
Session info ----------------------------------------------------------------------------------------
 system   x86_64, darwin15.6.0
 ui       RStudio (1.0.153)
 collate  en_US.UTF-8
 tz       America/Los_Angeles

Packages --------------------------------------------------------------------------------------------
 package      * version   date       source
 assertthat     0.2.0     2017-04-11 CRAN (R 3.4.1)
 auk          * 0.0.2.901 <NA>       local
 backports      1.1.0     2017-05-22 CRAN (R 3.4.1)
 base         * 3.4.1     2017-07-07 local
 bindr          0.1       2016-11-13 CRAN (R 3.4.1)
 bindrcpp     * 0.2       2017-06-17 CRAN (R 3.4.1)
 callr          1.0.0     2016-06-18 CRAN (R 3.4.0)
 clisymbols     1.2.0     2017-08-24 Github (gaborcsardi/clisymbols@e49b4f5)
 commonmark     1.2       2017-03-01 CRAN (R 3.4.1)
 compiler       3.4.1     2017-07-07 local
 countrycode    0.19      2017-02-06 CRAN (R 3.4.0)
 covr         * 3.0.0     2017-06-26 CRAN (R 3.4.1)
 crayon         1.3.2     2016-06-28 CRAN (R 3.4.1)
 cyclocomp      1.1.0     2017-08-24 Github (MangoTheCat/cyclocomp@6156a12)
 data.table     1.10.4    2017-02-01 CRAN (R 3.4.0)
 datasets     * 3.4.1     2017-07-07 local
 desc           1.1.1     2017-08-03 CRAN (R 3.4.1)
 devtools     * 1.13.3    2017-08-02 CRAN (R 3.4.1)
 digest         0.6.12    2017-01-27 CRAN (R 3.4.1)
 dplyr          0.7.2     2017-07-20 CRAN (R 3.4.1)
 evaluate       0.10.1    2017-06-24 CRAN (R 3.4.1)
 glue           1.1.1     2017-06-21 CRAN (R 3.4.1)
 goodpractice   1.0.0     2017-08-24 Github (MangoTheCat/goodpractice@9969799)
 graphics     * 3.4.1     2017-07-07 local
 grDevices    * 3.4.1     2017-07-07 local
 hms            0.3       2016-11-22 CRAN (R 3.4.0)
 httr           1.3.1     2017-08-20 CRAN (R 3.4.1)
 igraph         1.1.2     2017-07-21 CRAN (R 3.4.1)
 jsonlite       1.5       2017-06-01 CRAN (R 3.4.1)
 knitr          1.17      2017-08-10 CRAN (R 3.4.1)
 lazyeval       0.2.0     2016-06-12 CRAN (R 3.4.0)
 lintr          1.0.1     2017-08-10 CRAN (R 3.4.1)
 magrittr       1.5       2014-11-22 CRAN (R 3.4.1)
 memoise        1.1.0     2017-04-21 CRAN (R 3.4.1)
 methods      * 3.4.1     2017-07-07 local
 pkgconfig      2.0.1     2017-03-21 CRAN (R 3.4.1)
 praise         1.0.0     2015-08-11 CRAN (R 3.4.0)
 purrr          0.2.3     2017-08-02 CRAN (R 3.4.1)
 R6             2.2.2     2017-06-17 CRAN (R 3.4.1)
 rcmdcheck      1.2.1     2016-09-28 CRAN (R 3.4.0)
 Rcpp           0.12.12   2017-07-15 CRAN (R 3.4.1)
 readr          1.1.1     2017-05-16 CRAN (R 3.4.0)
 remotes        1.1.0     2017-07-09 CRAN (R 3.4.1)
 rex            1.1.1     2016-12-05 CRAN (R 3.4.0)
 rlang          0.1.2     2017-08-09 CRAN (R 3.4.1)
 roxygen2       6.0.1     2017-02-06 CRAN (R 3.4.1)
 rprojroot      1.2       2017-01-16 CRAN (R 3.4.1)
 rstudioapi     0.6       2016-06-27 CRAN (R 3.4.1)
 stats        * 3.4.1     2017-07-07 local
 stringi        1.1.5     2017-04-07 CRAN (R 3.4.1)
 stringr        1.2.0     2017-02-18 CRAN (R 3.4.1)
 testthat     * 1.0.2     2016-04-23 CRAN (R 3.4.0)
 tibble         1.3.4     2017-08-22 CRAN (R 3.4.1)
 tidyr          0.7.0     2017-08-16 CRAN (R 3.4.1)
 tools          3.4.1     2017-07-07 local
 utils        * 3.4.1     2017-07-07 local
 whoami         1.1.1     2015-07-13 CRAN (R 3.4.0)
 withr          2.0.0     2017-07-28 CRAN (R 3.4.1)
 xml2           1.1.1     2017-01-24 CRAN (R 3.4.1)
 xmlparsedata   1.0.1     2016-06-18 CRAN (R 3.4.0)
```
