### 2017-08-30 #136 @karthik
@aurielfournier Sorry I totally missed that your review was above Ted's. My apologies.
