### 2017-09-05 06b3e55
changes suggested by @aurielfournier for rOpenSci
