+++ inst/doc/auk.R
@@ -7,6 +7,33 @@ knitr::opts_chunk$set(
 suppressPackageStartupMessages(library(auk))
 suppressPackageStartupMessages(library(dplyr))
 
+## ----quickstart, eval = FALSE--------------------------------------------
+#  library(auk)
+#  library(dplyr)
+#  # path to the ebird data file, here a sample included in the package
+#  input_file <- system.file("extdata/ebd-sample.txt", package = "auk")
+#  # output text file
+#  output_file <- "ebd_filtered_grja.txt"
+#  ebird_data <- input_file %>%
+#    # 1. reference file
+#    auk_ebd() %>%
+#    # 2. define filters
+#    auk_species(species = "Gray Jay") %>%
+#    auk_country(country = "Canada") %>%
+#    # 3. run filtering
+#    auk_filter(file = output_file) %>%
+#    # 4. read text file into r data frame
+#    read_ebd()
+
+## ----quickstart-nopipes, eval = FALSE------------------------------------
+#  input_file <- system.file("extdata/ebd-sample.txt", package = "auk")
+#  output_file <- "ebd_filtered_grja.txt"
+#  ebd <- auk_ebd(input_file)
+#  ebd_filters <- auk_species(ebd, species = "Gray Jay")
+#  ebd_filters <- auk_country(ebd_filters, country = "Canada")
+#  ebd_filtered <- auk_filter(ebd_filters, file = output_file)
+#  ebd_df <- read_ebd(ebd_filtered)
+
 ## ----example-data-1, eval = FALSE----------------------------------------
 #  library(auk)
 #  library(dplyr)
@@ -25,8 +52,6 @@ suppressPackageStartupMessages(library(dplyr))
 #  f_out <- "ebd_cleaned.txt"
 #  # remove problem records
 #  cleaned <- auk_clean(f, f_out = f_out)
-#  # tidy up
-#  unlink(f_out)
 
 ## ----auk-ebd-------------------------------------------------------------
 ebd <- system.file("extdata/ebd-sample_messy.txt", package = "auk") %>% 
@@ -34,7 +59,7 @@ ebd <- system.file("extdata/ebd-sample_messy.txt", package = "auk") %>%
 ebd
 
 ## ----auk-filter----------------------------------------------------------
-ebd <- ebd %>% 
+ebd_filters <- ebd %>% 
   # species: common and scientific names can be mixed
   auk_species(species = c("Gray Jay", "Cyanocitta cristata")) %>%
   # country: codes and names can be mixed; case insensitive
@@ -49,17 +74,15 @@ ebd <- ebd %>%
   auk_duration(duration = c(0, 60)) %>%
   # complete: all species seen or heard are recorded
   auk_complete()
-ebd
+ebd_filters
 
 ## ----auk-complete, eval = FALSE------------------------------------------
 #  output_file <- "ebd_filtered_blja-grja.txt"
-#  ebd <- system.file("extdata/ebd-sample.txt", package = "auk") %>%
+#  ebd_jays <- system.file("extdata/ebd-sample.txt", package = "auk") %>%
 #    auk_ebd() %>%
 #    auk_species(species = c("Gray Jay", "Cyanocitta cristata")) %>%
 #    auk_country(country = "Canada") %>%
 #    auk_filter(file = output_file)
-#  # tidy up
-#  unlink(output_file)
 
 ## ----read----------------------------------------------------------------
 system.file("extdata/ebd-sample.txt", package = "auk") %>% 
@@ -68,14 +91,12 @@ system.file("extdata/ebd-sample.txt", package = "auk") %>%
 
 ## ----read-auk-ebd, eval = FALSE------------------------------------------
 #  output_file <- "ebd_filtered_blja-grja.txt"
[…]
