+++ R/auk-date.r
@@ -1,11 +1,11 @@
-#' Filter the EBD by date
+#' Filter the eBird data by date
 #'
 #' Define a filter for the eBird Basic Dataset (EBD) based on a range of dates.
 #' This function only defines the filter and, once all filters have been
 #' defined, [auk_filter()] should be used to call AWK and perform the
 #' filtering.
 #'
-#' @param x `auk_ebd` object; reference to EBD file created by [auk_ebd()].
+#' @param x `auk_ebd` object; reference to object created by [auk_ebd()].
 #' @param date character or date; date range to filter by , provided either as a
 #'   character vector in the format `"2015-12-31"` or a vector of Date objects.
 #'
