### 2017-09-05 #136 @mstrimas
Finally getting to this, here are responses to @aurielfournier comments:

- details later: I've put a "Quick Start" section at the top of the vignette that gives a simple example of how what the package does before delving into the details. I also put this section at the top of the readme. **Does this address what you were getting at?**
- remove `unlink`: done!
- don't use "EBD": replaced all instances of EBD with non-acronyms.
- reusing variable names: fixed, no repeatedly re-using same variable name anymore.
- use of pipes: I'm don't want every example in the readme and vignette to be duplicated, with and without pipes. So, I've just given one example at the top where I write the code both ways. **Do you think I should write all examples in function documentation with and without pipes?**
- no `auk_filter()` in function help: this was by design. Since `auk_filter()` is the only function in the package that has an external dependency (i.e. AWK) that isn't installed on some systems (e.g. Windows), I decided to use it minimally in the help examples. I can include it, but then I'll need to enclose all examples in `dontrun{}` blocks. Maybe this isn't a big deal though. **@emhart @aurielfournier, what are your thoughts, is having `auk_filter()` important enough warrant using `dontrun{}`?**

Thanks for the code review!!!
