### 2017-09-05 bb2362a
suggestions from @emhart for rOpenSci review
