+++ R/auk-filter.r
@@ -49,7 +49,7 @@
 #'   auk_country(country = c("US", "Canada")) %>%
 #'   auk_extent(extent = c(-100, 37, -80, 52)) %>%
 #'   auk_date(date = c("2012-01-01", "2012-12-31")) %>%
-#'   auk_time(time = c("06:00", "09:00")) %>%
+#'   auk_time(start_time = c("06:00", "09:00")) %>%
 #'   auk_duration(duration = c(0, 60)) %>%
 #'   auk_complete()
 #' \dontrun{
