### 2017-09-05 #136 @mstrimas
Here are my responses to @emhart:

- multi-step process: I agree, this is confusing, I've added a "Quick start" section at the top of the vignette and readme, which tries to explain this. **Do you think this clarifies things?**
- wrapper function: I think an all in one function is an interesting idea. I initially wanted to have a clear distinction between defining filters and running filters because running on the full dataset takes several hours. also, it's possible the filtered text file will be huge and not easily readable into R, which is why reading is also separated out. I'll look in to a wrapper function that does everything at once though, could be handy for users.
- user laziness: an interesting point that I hadn't thought of. I tried to allow users to be as lazy as possible without it being too hard to guess what they're doing. For dates, the user can provide a string (ISO format required) or a Date object. The latter option will allow users to user to first define dates in any format they like using `as.Date()`. **Do you have any thoughts on how to guess which format the text date is in?** Species are tough. "robin" doesn't work because there are so many different species of robin and no way to know you mean "american robin". More tricky is different common or scientific names for different species. Currently, for simplicity, the user must give the English common name or scientific name (case insensitive) as used in the eBird taxonomy. Ideally, we'd like to allow users to look up species using other taxonomies, names in different languages, alternate spellings, etc. but this is a daunting task that we probably won't get to for some time.
- extent units: yes, decimal degrees. fixed.
- cygwin on windows: **Do you think allowing users to specify an environment variable would solve this?**
- citation file: added.
- column filter: **do you mean a way to specify a subset of columns to write to the text file?** users can always dplyr::filter() after import, but I guess outputting fewer columns to the text file results in a smaller file size. I can think about this.
- time variable: oops, fixed this.
- CONTRIBUTING: I don't have a CONTRIBUTING file but I do have a CONDUCT file. **is this that same, should I rename this CONTRIBUTING?**

Thanks!!!
