+++ R/auk-getpath.r
@@ -1,6 +1,8 @@
 #' OS specific path to AWK
 #'
-#' Return the OS specific path to AWK, or highlights if it's not installed.
+#' Return the OS specific path to AWK, or highlights if it's not installed. To 
+#' manually set the path to AWK, set the `AWK_PATH` environment variable in your
+#' `.Renviron` file.
 #'
 #' @return Path to AWK or `NA` if AWK wasn't found.
 #' @export
@@ -9,9 +11,18 @@
 auk_getpath <- function() {
   sysname <- tolower(Sys.info()[["sysname"]])
 
-  # mac or linux
-  if (sysname %in% c("darwin", "linux")) {
-    # test
+  # manually specified path
+  if (Sys.getenv("AWK_PATH") != "") {
+    awk <- Sys.getenv("AWK_PATH")
+    awk_test <- tryCatch(
+      list(result = system(paste(awk, "--version"),
+                           intern = TRUE, ignore.stderr = TRUE)),
+      error = function(e) list(result = NULL),
+      warning = function(e) list(result = NULL)
+    )
+  } else if (sysname %in% c("darwin", "linux")) {
+    # mac or linux
+    # test and find path
     awk_test <- tryCatch(
       list(result = system("which awk", intern = TRUE, ignore.stderr = TRUE)),
       error = function(e) list(result = NULL),
