+++ README.md
@@ -93,7 +93,7 @@ f <- system.file("extdata/ebd-sample_messy.txt", package = "auk")
 tmp <- tempfile()
 # remove problem records
 auk_clean(f, tmp)
-#> [1] "/var/folders/mg/qh40qmqd7376xn8qxd6hm5lwjyy0h2/T//Rtmpyb1pPf/filedabb2004e523"
+#> [1] "/var/folders/mg/qh40qmqd7376xn8qxd6hm5lwjyy0h2/T//Rtmp97C8MK/file1165218f9812"
 # number of lines in input
 length(readLines(f))
 #> [1] 101
