### 2017-09-06 #136 @mstrimas
@emhart Just added ability to manually set awk path by setting the `AWK_PATH` environment variable in .Renviron. Should work on Mac or Windows, though I don't have a Windows machine to test.
