### 2017-09-07 d396199
fixed spaces in filenames causing error, closes issue #7
