+++ inst/doc/auk.Rmd
@@ -34,7 +34,6 @@ Because the eBird dataset is so large, step 3 typically takes several hours to r
 
 ```{r quickstart, eval = FALSE}
 library(auk)
-library(dplyr)
 # path to the ebird data file, here a sample included in the package
 input_file <- system.file("extdata/ebd-sample.txt", package = "auk")
 # output text file
