### 2017-09-11 #136 @aurielfournier
@mstrimas

The Quick Start is great, exactly what I was looking for.

I think one example is sufficient for the with and without pipes.

I see what you mean about `auk_filter()` now, I guess I could go either way on that one. Do you have thoughts @emhart ?
