### 2017-09-19 85b0c8b
fixed error in print method for extent filter
