### 2017-09-21 6a841e7
auk_distance added
