+++ NEWS.md
@@ -1,8 +1,12 @@
-# auk 0.0.2.900
+# auk 0.1.0
+
+- eBird taxonomy update to August 2017 version, users should download the most recent EBD to ensure the taxonomy is in sync with the new package
+- auk_distance filter now included to subset by distance travelled
+- Many changes suggested by rOpenSci package peer review process, see https://github.com/ropensci/onboarding/issues/136 for details
 
 # auk 0.0.2
 
-- patch release converting ebird_taxonomy to ASCII to pass CRAN checks
+- Patch release converting ebird_taxonomy to ASCII to pass CRAN checks
 
 # auk 0.0.1
 
