+++ man/auk_duration.Rd
@@ -19,11 +19,17 @@ An \code{auk_ebd} object.
 Define a filter for the eBird Basic Dataset (EBD) based on the duration of
 the checklist. This function only defines the filter and, once all filters
 have been defined, \code{\link[=auk_filter]{auk_filter()}} should be used to call AWK and perform the
-filtering.
+filtering. Note that checklists with no effort, such as incidental
+observations, will be excluded if this filter is used since they have no
+associated duration information.
 }
 \examples{
 # only keep checklists that are less than an hour long
 system.file("extdata/ebd-sample.txt", package = "auk") \%>\%
   auk_ebd() \%>\%
   auk_duration(duration = c(0, 60))
+  
+# alternatively, without pipes
+ebd <- auk_ebd(system.file("extdata/ebd-sample.txt", package = "auk"))
+auk_duration(ebd, duration = c(0, 60))
 }
