+++ docs/articles/awk-script.txt
@@ -26,6 +26,7 @@ BEGIN {
   
   
   
+  
 
   # keeps header
   if (NR == 1) {

[…]
