### 2017-09-26 #136 @emhart
Sorry for the delay @mstrimas here are a few quick thoughts:

* **Do you think this clarifies things?** - *yes*
* **Do you have any thoughts on how to guess which format the text date is in?** - *No, not really, honestly I'm impressed with the ease of use built in.  I would usually specify a format, and then do a check for the format and throw an error, I don't think you need to guess*
* **Do you think allowing users to specify an environment variable would solve this?** - *Sounds like the answer is yes, although I also don't have a windows machine*
* **do you mean a way to specify a subset of columns to write to the text file?** - *yes, that is what I meant, I guess I was thinking since you're writing it out to disk, instead of having the user read all the data in, why not let them just write the columns they want.  I think this falls under the "nice to have" features, not a necessary one*
* **is this that same, should I rename this CONTRIBUTING?** - *My understanding is that the CONTRIBUTING file gives guidelines for how users can contribute to the project . Here is an [example](https://github.com/ropensci/scrubr/blob/master/CONTRIBUTING.md) vs a [CONDUCT file](https://github.com/ropensci/scrubr/blob/master/CONDUCT.md)*
