### 2017-09-26 a3d6c6f
added non-pipe examples
