+++ R/auk-time.r
@@ -16,6 +16,10 @@
 #' system.file("extdata/ebd-sample.txt", package = "auk") %>%
 #'   auk_ebd() %>%
 #'   auk_time(start_time = c("06:00", "08:00"))
+#'   
+#' # alternatively, without pipes
+#' ebd <- auk_ebd(system.file("extdata/ebd-sample.txt", package = "auk"))
+#' auk_time(ebd, start_time = c("06:00", "08:00"))
 auk_time <- function(x, start_time)  {
   UseMethod("auk_time")
 }
