+++ R/auk-country.r
@@ -24,6 +24,10 @@
 #' system.file("extdata/ebd-sample.txt", package = "auk") %>%
 #'   auk_ebd() %>%
 #'   auk_country(country)
+#'   
+#' # alternatively, without pipes
+#' ebd <- auk_ebd(system.file("extdata/ebd-sample.txt", package = "auk"))
+#' auk_country(ebd, country)
 auk_country <- function(x, country, replace)  {
   UseMethod("auk_country")
 }
