+++ R/auk-species.r
@@ -20,6 +20,10 @@
 #' system.file("extdata/ebd-sample.txt", package = "auk") %>%
 #'   auk_ebd() %>%
 #'   auk_species(species)
+#'   
+#' # alternatively, without pipes
+#' ebd <- auk_ebd(system.file("extdata/ebd-sample.txt", package = "auk"))
+#' auk_species(ebd, species)
 auk_species <- function(x, species, replace)  {
   UseMethod("auk_species")
 }
