+++ NEWS.md
@@ -1,6 +1,7 @@
 # auk 0.1.0
 
 - eBird taxonomy update to August 2017 version, users should download the most recent EBD to ensure the taxonomy is in sync with the new package
+- Manually set AWK path with environment variable `AWK_PATH` in `.Renviron` file 
 - auk_distance filter now included to subset by distance travelled
 - Many changes suggested by rOpenSci package peer review process, see https://github.com/ropensci/onboarding/issues/136 for details
 
