### 2017-09-26 #136 @mstrimas
@emhart thanks! I like the idea of column subsetting and will start looking at the best way to implement that.
