### 2017-10-03 #136 @mstrimas
`auk_filter()` now has additional arguments `keep` and `drop` that users can use to specify which columns are output.
