### 2017-10-03 2afd44d
updating news for changes
