### 2017-10-12 2d41e80
added auk_project and auk_protocol
