+++ tests/testthat/test_filters.r
@@ -167,6 +167,45 @@ test_that("auk_last_edited", {
 
 })
 
+test_that("auk_protocol", {
+  ebd <- system.file("extdata/ebd-sample.txt", package = "auk") %>%
+    auk_ebd() %>% 
+    auk_protocol("stationary")
+  
+  # works correctly
+  expect_equal(ebd$filters$protocol, "stationary")
+  
+  # multiple protocols
+  ebd <- auk_protocol(ebd, c("stationary", "traveling"))
+  expect_equal(ebd$filters$protocol, c("stationary", "traveling"))
+  
+  # raises error for bad input
+  expect_error(auk_protocol(ebd, "STATIONARY"))
+  expect_error(auk_protocol(ebd, 2))
+  expect_error(auk_protocol(ebd, ""))
+  expect_error(auk_protocol(ebd, NA))
+})
+
+test_that("auk_project", {
+  ebd <- system.file("extdata/ebd-sample.txt", package = "auk") %>%
+    auk_ebd() %>% 
+    auk_project("EBIRD")
+  
+  # works correctly
+  expect_equal(ebd$filters$project, "EBIRD")
+  
+  # multiple projects
+  ebd <- auk_project(ebd, c("EBIRD", "EBIRD_MEX"))
+  expect_equal(ebd$filters$project, c("EBIRD", "EBIRD_MEX"))
+  
+  # raises error for bad input
+  expect_error(auk_project(ebd, "EBIRD MEX"))
+  expect_error(auk_project(ebd, "ebird_mex"))
+  expect_error(auk_project(ebd, 2))
+  expect_error(auk_project(ebd, ""))
+  expect_error(auk_project(ebd, NA))
+})
+
 test_that("auk_time", {
   ebd <- system.file("extdata/ebd-sample.txt", package = "auk") %>%
     auk_ebd()
