+++ docs/articles/awk-script.txt
@@ -27,6 +27,9 @@ BEGIN {
   
   
   
+  
+  
+  
 
   # keeps header
   if (NR == 1) {
