+++ vignettes/development.Rmd
@@ -9,9 +9,19 @@ editor_options:
   chunk_output_type: console
 ---
 
-This vignette describes the process of updating and extending `auk`. Two topics are covered: updating `auk` when a new eBird taxonomy is released and extending `auk` to include new filters.
+This vignette describes the process of updating and extending `auk`. Three topics are covered: updating `auk` when a new eBird taxonomy is released, extending `auk` to include new filters, and CRAN submission.
 
 ## Updating the eBird taxonomy
+
+The species, and other taxa, available for entry into the eBird database is dependent on the [eBird taxonomy](http://help.ebird.org/customer/portal/articles/1006825-the-ebird-taxonomy). Every August, the eBird team updates this taxonomy to reflect name changes splits, merges, new species, or any other changes. Historical eBird records are then updated accordingly and subsequent EBD files reflect this updated taxonomy. The `auk` package stores a copy of this taxonomy as the data frame `ebird_taxonomy`, and uses it both for filtering by species (`auk_species()`) and for taxonomic roll-up (`auk_rollup()`). Therefore, `auk` must be updated when a new eBird taxonomy is released. This section described how this is done. It is best to do this after the new taxonomy **and** the new EBD have both been released, otherwise the taxonomy and EBD will be out of sync.
+
+When the eBird taxonomy is updated, the new version can be downloaded from the [eBird website](http://help.ebird.org/customer/portal/articles/1006825-the-ebird-taxonomy). The taxonomy can be downloaded in csv or Excel format, **be sure to download the Excel file** because the csv file has character encoding issues. Copy this file to `data-raw/`. At this point, you should check that this new taxonomy has the same format as the previous file, which will also be in this directory. Ensure that the same columns are present and that they're named the same.
+
+The file `data-raw/ebird-taxonomy.r` prepares the taxonomy as a data frame to be stored in the package. Open this file and edit the `read_xlsx()` call to point to the new file you just downloaded. Run the code, then open the `ebird_taxonomy` data frame to inspect it and make sure there's no glaring issues. One potential error that should be investigated is non-ASCII characters. Some common names have accented characters (e.g. Rüppell's Griffon, Gyps rueppelli), which can cause problems. `ebird-taxonomy.r` converts these characters to their unaccented equivalents (e.g. Ruppell's Griffon). Check that this record, or others with accented characters, has been properly converted.
+
+Next, update `auk_version_date()` (`R/auk-version-date.r`) to reflect the date of the new taxonomy and the new EBD.
+
+Finally, build the package (`devtools::build()`) and run `R CMD check` (`devtools::check()`). If everything looks good, commit to git and push to GitHub.
 
 ## Adding new filters
 
@@ -21,12 +31,12 @@ The primary functionality of `auk` is to apply filters to the EBD to extract a s
 2. Create a new function `auk_color()` (in file `R/auk-color.r`) that defines the new filter. As a starting point, use one of the other filtering functions. For example to filter on a range of numeric values, start with `auk_duration()`, to filter on a logical (true/false) variable use `auk_complete()`, or to filter on a discrete, categorical variable use `auk_country()`. Be sure to apply extensive checking on the validity of inputs and update the documentation, including examples.
 3. Update `auk_filter()` (in file `R/auk-filter.r`) to incorporate the filtering criteria into the AWK script. Again, use an existing filter as a template.
 4. Create unit tests for the new filter by creating a new `test_that()` block in `tests/testthat/test_filters.r`. Again, use an existing filter as a template.
-5. Update `README.md` and `vignettes/auk.Rmd` to add the new filter to the list of potential filtes.
+5. Update `README.md` and `vignettes/auk.Rmd` to add the new filter to the list of potential filters.
 6. Build, test, check, and push to GitHub
 
 ### 1. Update `auk_ebd()`
 
[…]
