+++ tests/testthat/test_auk-filter.r
@@ -38,6 +38,26 @@ test_that("auk_filter filter an ebd", {
   expect_true(all(ebd$longitude <= filters$filters$extent[3]))
   expect_true(all(ebd$latitude >= filters$filters$extent[2]))
   expect_true(all(ebd$latitude <= filters$filters$extent[4]))
+  
+  # filter again
+  tmp <- tempfile()
+  ebd <- auk_ebd(f) %>%
+    auk_project("EBIRD_CAN") %>% 
+    auk_protocol("stationary") %>% 
+    auk_filter(file = tmp) %>% 
+    read_ebd()
+  unlink(tmp)
+  
+  expect_true(all(ebd$project_code == "EBIRD_CAN"))
+  expect_true(all(ebd$protocol_type == "eBird - Stationary Count"))
+  
+  # again
+  ebd <- system.file("extdata/zerofill-ex_ebd.txt", package = "auk") %>% 
+    auk_ebd() %>%
+    auk_breeding() %>% 
+    auk_filter(file = tmp) %>% 
+    read_ebd()
+  expect_true(all(!is.na(ebd$breeding_bird_atlas_code)))
 })
 
 test_that("auk_filter filter sampling and ebd files", {
