+++ vignettes/development.Rmd
@@ -38,7 +38,9 @@ The primary functionality of `auk` is to apply filters to the EBD to extract a s
 
 Near the top of the code for `auk_ebd()`, a data frame named `filter_cols` is defined which specifies which columns have associated filters. Add a new row to this data frame and set `name` as the name of the column in the file header that will be filtered on and `id` as the name of the filter. For example, if you're creating a filter called `auk_color()` that filters on the column "FEATHER COLOR", then set `id = "color"` and `name = "feather color"`. Ideally, similar filters should be grouped together in this data frame, so insert the new row accordingly.
 
-Next, at the end of the code for `auk_ebd()`, the `auk_ebd` object is created and returned with the statement beginning with `structure(...`. This object should have placeholders for every filter. So, add a new element to the list, naming the variable after the `id` in the above data frame, putting it in the same order as in the above data frame, and choosing a sensible data type. For example, if `color` is a categorical variable, add a new list element `color = character()`, and if it is a logical (T/F) variable, add `color = logical()`.
+For filters that don't apply to the sampling event data file, i.e. filters at the species level rather than the checklist level, add the id to the character vector `not_in_sampling`. For example, modify the code to read: `not_in_sampling <- c("species", "breeding", "color")`.
+
+Next, at the end of the code for `auk_ebd()`, the `auk_ebd` object is created and returned with the statement beginning with `structure(...`. This object should have placeholders for every filter. So, add a new element to the list, naming the variable after the `id` in the above data frame, putting it in the same order as in the above data frame, and choosing a sensible data type. For example, if `color` is a categorical variable, add a new list element `color = character()`, and if it's a numeric variable, add `color = numeric()`.
 
 Finally, within `auk-ebd.r` a `print.auk_ebd()` method is defined, which you'll need to update to print the filter in a sensible way. Here you're best to find another filter with a similar format and use that as a template. Again, be sure to put the print code for the filter in the right order. For example, for a categorical filter allow multiple potential values, you may way something like:
 
