+++ man/auk_breeding.Rd
@@ -1,0 +1,27 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/auk-breeding.r
+\name{auk_breeding}
+\alias{auk_breeding}
+\title{Filter to only inlcude observations with breeding codes}
+\usage{
+auk_breeding(x)
+}
+\arguments{
+\item{x}{\code{auk_ebd} object; reference to basic dataset file created by
+\code{\link[=auk_ebd]{auk_ebd()}}.}
+}
+\value{
+An \code{auk_ebd} object.
+}
+\description{
+eBird users have the option of specifying breeding bird atlas codes for their
+observations, for example, if nesting building behaviour is observered. Use
+this filter to select only those observations with an associated breeding
+code. This function only defines the filter and, once all filters have been
+defined, \code{\link[=auk_filter]{auk_filter()}} should be used to call AWK and perform the filtering.
+}
+\examples{
+system.file("extdata/ebd-sample.txt", package = "auk") \%>\%
+  auk_ebd() \%>\%
+  auk_breeding()
+}

[…]
