+++ cran-comments.md
@@ -8,7 +8,7 @@
 
 # R CMD check results
 
-There were no ERROR or WARNINGs.
+There were no ERRORs or WARNINGs.
 
 There was 1 NOTE:
 
