### 2017-10-20 85df493
not correctly defaulting to readr reading
