+++ docs/articles/auk.R
@@ -1,0 +1,191 @@
+## ---- echo = FALSE-------------------------------------------------------
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>",
+  warning = FALSE, error = FALSE, message = FALSE
+)
+suppressPackageStartupMessages(library(auk))
+suppressPackageStartupMessages(library(dplyr))
+
+## ----quickstart, eval = FALSE--------------------------------------------
+#  library(auk)
+#  # path to the ebird data file, here a sample included in the package
+#  input_file <- system.file("extdata/ebd-sample.txt", package = "auk")
+#  # output text file
+#  output_file <- "ebd_filtered_grja.txt"
+#  ebird_data <- input_file %>%
+#    # 1. reference file
+#    auk_ebd() %>%
+#    # 2. define filters
+#    auk_species(species = "Gray Jay") %>%
+#    auk_country(country = "Canada") %>%
+#    # 3. run filtering
+#    auk_filter(file = output_file) %>%
+#    # 4. read text file into r data frame
+#    read_ebd()
+
+## ----quickstart-nopipes, eval = FALSE------------------------------------
+#  input_file <- system.file("extdata/ebd-sample.txt", package = "auk")
+#  output_file <- "ebd_filtered_grja.txt"
+#  ebd <- auk_ebd(input_file)
+#  ebd_filters <- auk_species(ebd, species = "Gray Jay")
+#  ebd_filters <- auk_country(ebd_filters, country = "Canada")
+#  ebd_filtered <- auk_filter(ebd_filters, file = output_file)
+#  ebd_df <- read_ebd(ebd_filtered)
+
+## ----example-data-1, eval = FALSE----------------------------------------
+#  library(auk)
+#  library(dplyr)
+#  system.file("extdata/ebd-sample.txt", package = "auk")
+
+## ----example-data-2, eval = FALSE----------------------------------------
+#  # ebd
+#  system.file("extdata/zerofill-ex_ebd.txt", package = "auk")
+#  # sampling event data
+#  system.file("extdata/zerofill-ex_sampling.txt", package = "auk")
+
+## ----auk-clean, eval = FALSE---------------------------------------------
+#  library(auk)
+#  # sample data, with intentially introduced errors
+#  f <- system.file("extdata/ebd-sample_messy.txt", package = "auk")
+#  f_out <- "ebd_cleaned.txt"
+#  # remove problem records
+#  cleaned <- auk_clean(f, f_out = f_out)
+
+## ----auk-ebd-------------------------------------------------------------
+ebd <- system.file("extdata/ebd-sample_messy.txt", package = "auk") %>% 
+  auk_ebd()
+ebd
+
+## ----auk-filter----------------------------------------------------------
+ebd_filters <- ebd %>% 
+  # species: common and scientific names can be mixed
+  auk_species(species = c("Gray Jay", "Cyanocitta cristata")) %>%
+  # country: codes and names can be mixed; case insensitive
+  auk_country(country = c("US", "Canada", "mexico")) %>%
+  # extent: long and lat in decimal degrees
+  # formatted as `c(lng_min, lat_min, lng_max, lat_max)`
+  auk_extent(extent = c(-100, 37, -80, 52)) %>%
+  # date: use standard ISO date format `"YYYY-MM-DD"`
+  auk_date(date = c("2012-01-01", "2012-12-31")) %>%
+  # time: 24h format
+  auk_time(start_time = c("06:00", "09:00")) %>%
+  # duration: length in minutes of checklists
+  auk_duration(duration = c(0, 60)) %>%
+  # complete: all species seen or heard are recorded
+  auk_complete()
+ebd_filters
+
[…]
