### 2017-10-20 2b6196d
formatting refernces in vignette
