### 2017-10-20 419e640
bump test version in cran-comments.md
