+++ R/auk-zerofill.r
@@ -1,7 +1,7 @@
 #' Read and zero-fill an eBird data file
 #'
-#' Read an eBird Basic Dataset file, and associated sampling event data file, to
-#' produce a zero-filled, presence-absence dataset. The basic dataa contains bird
+#' Read an eBird Basic Dataset (EBD) file, and associated sampling event data file, to
+#' produce a zero-filled, presence-absence dataset. The EBD contains bird
 #' sightings and the sampling event data is a set of all checklists, they can be
 #' combined to infer absence data by assuming any species not reported on a
 #' checklist was had a count of zero.
