+++ R/auk-breeding.r
@@ -1,7 +1,7 @@
-#' Filter to only inlcude observations with breeding codes
+#' Filter to only include observations with breeding codes
 #'
 #' eBird users have the option of specifying breeding bird atlas codes for their
-#' observations, for example, if nesting building behaviour is observered. Use
+#' observations, for example, if nesting building behaviour is observed. Use
 #' this filter to select only those observations with an associated breeding
 #' code. This function only defines the filter and, once all filters have been
 #' defined, [auk_filter()] should be used to call AWK and perform the filtering.
