### 2017-10-21 52b7434
removed license file
