### 2017-10-21 90ee9cf
prepare for new release after cran
