+++ NEWS.md
@@ -1,3 +1,5 @@
+# auk 0.1.0.900
+
 # auk 0.1.0
 
 - eBird taxonomy update to August 2017 version, users should download the most recent EBD to ensure the taxonomy is in sync with the new package
