+++ R/auk-ebd.r
@@ -53,6 +53,7 @@ auk_ebd <- function(file, file_sampling, sep = "\t") {
 
   # read header rows
   header <- tolower(get_header(file, sep))
+  header <- stringr::str_replace_all(header, "_", " ")
   col_idx <- data.frame(id = NA_character_, 
                         name = header, 
                         index = seq_along(header),
