### 2018-01-22 #136 @aurielfournier
Oh crap. I'm so sorry. I'll do my best to get to this as soon as I can, but the federal shutdown is messing up my week majorly, and even if they do come back tomorrow its going to take a few days for things to even out again.
