### 2018-01-25 #136 @emhart
I took a look @mstrimas happy to sign off as well!  Thanks for incorporating the feedback, the package looks great.
