### 2018-01-30 #136 @stefaniebutland
Hello @mstrimas. Congratulations on `auk` acceptance! We would love to host a post about it, so if you're interested, have a look at the editorial and technical info here https://github.com/ropensci/roweb2#contributing-a-blog-post and let me know if you are considering it.
