### 2018-01-31 #136 @mstrimas
@stefaniebutland sure, I'm happy to modify the [vignette](https://cornelllabofornithology.github.io/auk/articles/auk.html) into a blog post. Probably won't be able to get to it for a week or two though.
