### 2018-02-28 62d5fe8
auk_select() and ability to work with sampling events only added
