+++ R/auk-date.r
@@ -5,9 +5,13 @@
 #' defined, [auk_filter()] should be used to call AWK and perform the
 #' filtering.
 #'
-#' @param x `auk_ebd` object; reference to object created by [auk_ebd()].
+#' @param x `auk_ebd` or `auk_sampling` object; reference to file created by 
+#'   [auk_ebd()] or [auk_sampling()].
 #' @param date character or date; date range to filter by, provided either as a
 #'   character vector in the format `"2015-12-31"` or a vector of Date objects.
+#' 
+#' @details This function can also work with on an `auk_sampling` object if the 
+#'   user only wishes to filter the sampling event data.
 #'
 #' @return An `auk_ebd` object.
 #' @export
@@ -46,4 +50,9 @@ auk_date.auk_ebd <- function(x, date) {
   # define filter
   x$filters$date <- date
   return(x)
+}
+
+#' @export
+auk_date.auk_sampling <- function(x, date) {
+  auk_date.auk_ebd(x, date)
 }
