+++ NAMESPACE
@@ -57,6 +57,7 @@ export(auk_rollup)
 export(auk_sampling)
 export(auk_select)
 export(auk_species)
+export(auk_split)
 export(auk_time)
 export(auk_unique)
 export(auk_version_date)

[…]
