### 2018-03-19 5fbafdc
cleaned up docs
