+++ R/auk-rollup.r
@@ -1,29 +1,61 @@
 #' Roll up eBird taxonomy to species
 #' 
-#' The eBird Basic Dataset (EBD) includes both true species and other taxa,  
-#' including domestics, hybrids, subspecies, "spuhs", and recognizable forms. 
-#' In some cases, a checklist may contain multiple records for the same species, 
-#' for example, both Audubon and Myrtle Yellow-rumped Warblers, as well as some 
-#' records that are not resolvable to species, for example, "warbler sp.". 
-#' This function addresses these cases by removing taxa not identifiable to 
-#' species and rolling up taxa identified below species level to a single record 
-#' for each species in each checklist. By default, [read_ebd()] calls 
-#' `ebd_rollup()` when importing an eBird data file.
+#' The eBird Basic Dataset (EBD) includes both true species and every other
+#' field-identifiable taxon that could be relevant for birders to report. This 
+#' includes taxa not identifiable to a species (e.g. hybrids) and taxa reported
+#' below the species level (e.g. subspecies). This function produces a list of 
+#' observations of true species, by removing the former and rolling the latter 
+#' up to the species level. By default, [read_ebd()] calls `ebd_rollup()` when 
+#' importing an eBird data file.
 #'
 #' @param x data.frame; data frame of eBird data, typically as imported by
 #'   [read_ebd()]
+#'   
 #' @details When rolling observations up to species level the observed counts
 #'   are summed across any taxa that resolve to the same species. However, if
 #'   any of these taxa have a count of "X" (i.e. the observer did not enter a
 #'   count), then the rolled up record will get an "X" as well. For example, if 
-#'   an observer saw 3 Myrtle and 2 Audubon Warblers, this will roll up to 5 
+#'   an observer saw 3 Myrtle and 2 Audubon's Warblers, this will roll up to 5 
 #'   Yellow-rumped Warblers. However, if an "X" was entered for Myrtle, this 
 #'   would roll up to "X" for Yellow-rumped Warbler.
+#'   
+#'   The eBird taxonomy groups taxa into eight different categories. These 
+#'   categories, and the way they are treated by [auk_rollup()] are as follows:
+#'   
+#'   - **Species:** e.g., Mallard. Combined with lower level taxa if present on 
+#'   the same checklist.
+#'   - **ISSF or Identifiable Sub-specific Group:** Identifiable subspecies or
+#'   group of subspecies, e.g., Mallard (Mexican). If the checklist contains 
+#'   multiple taxa corresponding to the same species, the lower level taxa are 
+#'   rolled up, otherwise these records are left as is
+#'   - **Form:** Miscellaneous other taxa, including recently-described species
+#'   yet to be accepted or distinctive forms that are not universally accepted
+#'   (Red-tailed Hawk (Northern), Upland Goose (Bar-breasted)). If the checklist
+#'   contains multiple taxa corresponding to the same species, the lower level
+#'   taxa are rolled up, otherwise these records are left as is.
+#'   - **Spuh:**  Genus or identification at broad level -- e.g., duck sp.,
+#'   dabbling duck sp.. Dropped by auk_rollup().
+#'   - **Slash:** Identification to Species-pair e.g., American Black
+#'   Duck/Mallard). Dropped by auk_rollup()
+#'   - **Hybrid:** Hybrid between two species, e.g., American Black Duck x
+#'   Mallard (hybrid). Dropped by auk_rollup()
+#'   - **Intergrade:** Hybrid between two ISSF (subspecies or subspecies groups),
[…]
