+++ R/auk-ebd.r
@@ -38,7 +38,8 @@
 #'   filters once created with other package functions.
 #' @export
 #' @examples
-#' # set up reference to example file
+#' # get the path to the example data included in the package
+#' # in practice, provide path to ebd, e.g. f <- "data/ebd_relFeb-2018.txt
 #' f <- system.file("extdata/ebd-sample.txt", package = "auk")
 #' auk_ebd(f)
 #' # to produce zero-filled data, provide a checklist file
