### 2018-03-19 0507ca5
rollup now drops subspecies cols and sets category to species; rollup and unique set as attributes closing issue #4
