### 2018-03-19 8fd83aa
arrays implemented to speed up filtering
