+++ man/auk_protocol.Rd
@@ -12,7 +12,7 @@ auk_protocol(x, protocol)
 
 \item{protocol}{character; "Stationary", "Traveling", or "Incidental". Other
 protocols exist in the database, however, this function only extracts these
-three standard protocols.}
+three standard protocols. Multiple protocols are accepted.}
 }
 \value{
 An \code{auk_ebd} object.
