### 2018-03-19 2ce5cbf
using awk arrays to speed up filters
