+++ _pkgdown.yml
@@ -6,9 +6,6 @@ navbar:
     href: articles/auk.html
   - text: Function Reference
     href: reference/index.html
-  right:
-  - icon: fa-github fa-lg
-    href: https://github.com/CornellLabofOrnithology/auk
 
 authors:
   Matthew Strimas-Mackey:
