### 2018-03-23 49d32f2
removed all non-ASCII characters from example files, closes #14
