+++ data-raw/ebd-samples.r
@@ -1,30 +1,32 @@
 library(auk)
 library(tidyverse)
+library(stringi)
+library(stringr)
 
 ebird_dir <- "~/data/ebird/ebd_relFeb-2018/"
 # ebd
-f <- file.path(ebird_dir, "ebd_relFeb-2018.txt")
+f_in <- file.path(ebird_dir, "ebd_relFeb-2018.txt")
 f_subset <- file.path(ebird_dir, "ebd_relFeb-2018_subset.txt")
 f_sg <- file.path(ebird_dir, "ebd_relFeb-2018_SG.txt")
 f_ru <- file.path(ebird_dir, "ebd_relFeb-2018_rollup.txt")
 # sampling
-s <- file.path(ebird_dir, "ebd_sampling_relFeb-2018.txt")
+s_in <- file.path(ebird_dir, "ebd_sampling_relFeb-2018.txt")
 s_subset <- file.path(ebird_dir, "ebd_sampling_relFeb-2018_subset.txt")
 s_sg <- file.path(ebird_dir, "ebd_sampling_relFeb-2018_SG.txt")
 
 # filter
-filters <- auk_ebd(f) %>%
+filters <- auk_ebd(f_in) %>%
   auk_species(species = c("Gray Jay", "Blue Jay", "Steller's Jay", "Green Jay")) %>%
   auk_country(country = c("US", "Canada", "Mexico", "Belize", "Guatemala", "Honduras", "Panama", "Costa Rica", "El Salvador")) %>%
   auk_date(date = c("2010-01-01", "2012-12-31")) %>%
   auk_time(start_time = c("06:00", "12:00")) %>%
   auk_duration(duration = c(0, 120))
-filter_time <- system.time({
+if (!file.exists(f_subset)) {
   auk_filter(filters, file = f_subset, overwrite = TRUE)
-})
+}
 
-x <- read_tsv(f_subset, quote = "", col_types = cols(.default = col_character())) %>% 
-  mutate(`LAST NAME` = "eBirder") %>% 
+x <- read_tsv(f_subset, quote = "", 
+              col_types = cols(.default = col_character())) %>% 
   select(-X49)
 # evenly sample species
 set.seed(1)
@@ -36,53 +38,108 @@ x <- x %>%
 # sample to 500 records, make sure to get some from central america
 y1 <- sample_n(x %>% filter(!`COUNTRY CODE` %in% c("CA", "US")), 100)
 y2 <- sample_n(x %>% filter(`COUNTRY CODE` %in% c("CA", "US")), 400)
-y <- bind_rows(y1, y2)
-write_tsv(y %>% select(-X49), "inst/extdata/ebd-sample.txt", na = "")
+y <- bind_rows(y1, y2) %>% 
+  # fill with fake names for privacy
+  mutate(`FIRST NAME` = stri_trans_general(`FIRST NAME`, "latin-ascii"),
+         `LAST NAME` = "eBirder")
+f <- "inst/extdata/ebd-sample.txt"
+write_tsv(y, f, na = "")
+# remove any non-ascii characters
+readLines(f) %>% 
+  stri_trans_general("latin-ascii") %>% 
+  iconv("latin1", "ASCII", sub="") %>% 
+  str_replace_all("'|\"", "") %>% 
+  writeLines(f)
+stopifnot(length(tools::showNonASCII(readLines(f))) == 0)
+stopifnot(all(read_ebd(f)$scientific_name %in% ebird_taxonomy$scientific_name))
 # prepare a smaller sample of messy data
+y <- read_tsv(f, quote = "", 
+              col_types = cols(.default = col_character()))
+y$empty_col <- NA_character_
 names(y)[length(y)] <- ""
-y <- sample_n(y, 100)
[…]
