### 2018-03-28 c046aa6
state filtering fixed, closes issue #16
