+++ inst/extdata/ebd-sample_messy.txt
@@ -1,22 +1,22 @@
 GLOBAL UNIQUE IDENTIFIER	LAST EDITED DATE	TAXONOMIC ORDER	CATEGORY	COMMON NAME	SCIENTIFIC NAME	SUBSPECIES COMMON NAME	SUBSPECIES SCIENTIFIC NAME	OBSERVATION COUNT	BREEDING BIRD ATLAS CODE	BREEDING BIRD ATLAS CATEGORY	AGE/SEX	COUNTRY	COUNTRY CODE	STATE	STATE CODE	COUNTY	COUNTY CODE	IBA CODE	BCR CODE	USFWS CODE	ATLAS BLOCK	LOCALITY	LOCALITY ID	LOCALITY TYPE	LATITUDE	LONGITUDE	OBSERVATION DATE	TIME OBSERVATIONS STARTED	OBSERVER ID	FIRST NAME	LAST NAME	SAMPLING EVENT IDENTIFIER	PROTOCOL TYPE	PROTOCOL CODE	PROJECT CODE	DURATION MINUTES	EFFORT DISTANCE KM	EFFORT AREA HA	NUMBER OBSERVERS	ALL SPECIES REPORTED	GROUP IDENTIFIER	HAS MEDIA	APPROVED	REVIEWED	REASON	TRIP COMMENTS	SPECIES COMMENTS	
 URN:CornellLabOfOrnithology:EBIRD:OBS528997473	2017-08-31 02:48:51	18772	species	Green Jay	Cyanocorax yncas			1				Mexico	MX	Chiapas	MX-CHP			MX_169	60			07_020_PDLLB	L6229435	P	15.4712848	-92.545146	2011-09-07	10:20:00	obsr313215	MONITORES COMUNITARIOS	eBirder	S38916072	Stationary	P21	EBIRD	10			1	1		0	1	0		Agustn Mndez Lpez		
-URN:CornellLabOfOrnithology:EBIRD:OBS139325082	2018-02-13 20:47:26	18772	species	Green Jay	Cyanocorax yncas			10				United States	US	Texas	US-TX	Starr	US-TX-427		36			Salineno Wildlife Preserve--DeWinds Yard	L287918	H	26.514616	-99.1155021	2012-02-14	10:30:00	obsr118853	Robert	eBirder	S9836180	Traveling	P22	EBIRD	70	8.047		2	1		0	1	0				
-URN:CornellLabOfOrnithology:EBIRD:OBS116558645	2011-04-28 10:15:03	18816	species	Stellers Jay	Cyanocitta stelleri			2				United States	US	Washington	US-WA	Pierce	US-WA-053		5			My Backyard	L663043	P	47.2377151	-122.5431092	2011-04-27	06:15:00	obsr175451	Gary	eBirder	S8136321	Area	P23	EBIRD	35		.4047	1	1		0	1	0				
+URN:CornellLabOfOrnithology:EBIRD:OBS139325082	2018-02-13 20:47:26	18772	species	Green Jay	Cyanocorax yncas			10				United States	US	Texas	US-TX	Starr	US-TX-427		36			Salineno Wildlife Preserve--DeWind's Yard	L287918	H	26.514616	-99.1155021	2012-02-14	10:30:00	obsr118853	Robert	eBirder	S9836180	Traveling	P22	EBIRD	70	8.047		2	1		0	1	0				
[…]
