+++ README.md
@@ -138,7 +138,7 @@ drops these erroneous records and removes the blank column.
     f_out <- tempfile()
     # remove problem records
     auk_clean(f, f_out)
-    #> [1] "/var/folders/mg/qh40qmqd7376xn8qxd6hm5lwjyy0h2/T//RtmpsecHbt/file7fcc25c943e1"
+    #> [1] "/var/folders/mg/qh40qmqd7376xn8qxd6hm5lwjyy0h2/T//RtmpW4B7dt/file956167ef8afd"
     # number of lines in input
     length(readLines(f))
     #> [1] 51

[…]
