### 2018-04-13 CornellLabofOrnithology/auk#17 @maelle
Wrong review badge
It doesn't refer to the right onboarding thread 😉

It should be
