+++ README.md
@@ -138,13 +138,13 @@ drops these erroneous records and removes the blank column.
     f_out <- tempfile()
     # remove problem records
     auk_clean(f, f_out)
-    #> [1] "/var/folders/mg/qh40qmqd7376xn8qxd6hm5lwjyy0h2/T//RtmpW4B7dt/file956167ef8afd"
+    #> [1] "/var/folders/mg/qh40qmqd7376xn8qxd6hm5lwjyy0h2/T//Rtmpo2oPm7/file1fb862cf5c73"
     # number of lines in input
     length(readLines(f))
     #> [1] 51
     # number of lines in output
     length(readLines(f_out))
-    #> [1] 46
+    #> [1] 47
 
 ### Filtering
 
@@ -212,7 +212,7 @@ defined, the filtering is actually conducted using `auk_filter()`.
       auk_complete()
     ebd
     #> Input 
-    #>   EBD: /Library/Frameworks/R.framework/Versions/3.4/Resources/library/auk/extdata/ebd-sample.txt 
+    #>   EBD: /Users/mes335/projects/auk/inst/extdata/ebd-sample.txt 
     #> 
     #> Output 
     #>   Filters not executed
