### 2018-04-18 #136 @stefaniebutland
@mstrimas
> looks like there's going to be a large update to the data underlying this package in mid March

Checking in to see if timing is right for to draft a blog post - no rush if pkg not updated yet
