### 2018-04-18 #136 @maelle
By the way, it'd be grand if the blog post explained a bit how to choose between using `auk` and @sckott's `rebird` depending on the use case. 😺 The information could also be in the READMEs of both packages. Thinking of this because this week I was at a loss which of the two to recommend. 😀
