### 2018-04-18 #136 @mstrimas
@stefaniebutland I think the package is ready, I can start putting something together this week. Thanks for the reminder!

@maelle `rebird` is an interface to the [eBird API](https://confluence.cornell.edu/display/CLOISAPI/eBirdAPIs), which gives access to a very limited subset of the data, e.g. the last 30 days of observations from a location. I think of `rebird` as being useful for building tools and visualizations for birders; however, for most ecological applications (e.g. distribution modeling) you'll want access to the full eBird database (~500 million records).
