### 2018-04-18 #136 @mstrimas
@maelle I wasn't aware of the `ebirdfreq()` function, that's cool! Seems all the other functions are "recent" observations, but that one does give access to historical data at state, county, and hotspot level. It's also worth noting the `rebird` is easier to use and *much* faster, so if your data needs can be met by `rebird`, I'd say it's definitely preferred.
