### 2018-04-18 bdf370a
clarifying the difference between auk and rebird, suggested by @maelle
