+++ README.Rmd
@@ -45,6 +45,10 @@ devtools::install_github("CornellLabofOrnithology/auk")
 
 Full details on using `auk` to produce both presence-only and presence-absence data are outlined in the vignette, which can be accessed with `vignette("auk")`.
 
+## `auk` vs. `rebird`
+
+Those interested in eBird data may also want to consider [`rebird`](https://github.com/ropensci/rebird), an R package that provides an interface to the [eBird APIs](https://confluence.cornell.edu/display/CLOISAPI/eBirdAPIs). The functions in `rebird` are mostly limited to accessing recent (i.e. within the last 30 days) observations, although `ebirdfreq()` does provide historical frequency of observation data. In contrast, `auk` gives access to the full set of ~ 500 million eBird observations. For most ecological applications, users will require `auk`; however, for some use cases, e.g. building tools for birders, `rebird` provides a quick and easy way to access data.
+
 ## A note on versions
 
 This package contains a current (as of the time of package release) version of the [bird taxonomy used by eBird](http://help.ebird.org/customer/portal/articles/1006825-the-ebird-taxonomy). This taxonomy determines the species that can be reported in eBird and therefore the species that users of `auk` can extract. eBird releases an updated taxonomy once a year, typically in August, at which time `auk` will be updated to include the current taxonomy. When using `auk`, users should be careful to ensure that the version they're using is in sync with the eBird Basic Dataset they're working with. This is most easily accomplished by always using the must recent version of `auk` and the most recent release of the dataset.
