### 2018-04-18 #136 @maelle
Fantastic! Speaking of other rOpenSci packages, I am also wondering whether/how one could use [`bowerbird`](https://github.com/ropensci/bowerbird) (not an ornithology package despite its name) and `auk` to keep, update and use a local copy of eBird dataset, I might ping you if I ever try to write such an use case.
