### 2018-04-23 #136 @mstrimas
@stefaniebutland here's a first draft of a [blog post](https://github.com/mstrimas/auk-blog-post/blob/master/auk.md).

What topicid and date should I use?  Also, is there somewhere in the website repo I can put a couple data files (~ 3 MB). If there isn't a good spot, I'll just leave them in my GitHub repo.

If this looks good I can submit a pull request to the rOpenSci website repo.
