+++ makefile.r
@@ -16,3 +16,5 @@ rmarkdown::render("README.Rmd")
 devtools::build_vignettes()
 pkgdown::build_site()
 file.copy(list.files(".", "README.*png", full.names = TRUE), "docs/")
+dir.create("docs/hex-logo/")
+file.copy("hex-logo/auk.svg", "docs/hex-logo/")
