### 2018-05-14 #136 @stefaniebutland
@mstrimas I apologize for my long delay in responding. I am temporarily putting most blog post reviews and scheduling on hold until after our unconference. Rest assured your post will get a proper review and be published on our blog. I anticipate getting to this in early June.

> What topicid and date should I use?

Leave topicid blank. I add that immediately before publication and that links comments to our [discussion forum](https://discuss.ropensci.org/). We'll determine the date when I review your draft. Sorry about this delay. We really appreciate it when package authors do the extra work of contributing a post!
