### 2018-06-26 #136 @stefaniebutland
Hi @mstrimas. I hope you're still up for a post about `auk`. Right now I have a spot open for publication on 2018-07-24 so you could use that in your draft and if an earlier spot opens we can publish sooner and change the date.

> is there somewhere in the website repo I can put a couple data files (~ 3 MB). If there isn't a good spot, I'll just leave them in my GitHub repo.

Scott says good to leave them in your GitHub repo.

Happy to answer any questions
