### 2018-06-27 3da3c7b
changes for v1.9 of the EBD
