+++ data-raw/ebd-samples.r
@@ -3,21 +3,24 @@ library(tidyverse)
 library(stringi)
 library(stringr)
 
-ebird_dir <- "~/data/ebird/ebd_relFeb-2018/"
+ebird_dir <- "~/data/ebird/ebd_relMay-2018/"
 # ebd
-f_in <- file.path(ebird_dir, "ebd_relFeb-2018.txt")
-f_subset <- file.path(ebird_dir, "ebd_relFeb-2018_subset.txt")
-f_sg <- file.path(ebird_dir, "ebd_relFeb-2018_SG.txt")
-f_ru <- file.path(ebird_dir, "ebd_relFeb-2018_rollup.txt")
+f_in <- file.path(ebird_dir, "ebd_relMay-2018.txt")
+f_subset <- file.path(ebird_dir, "ebd_relMay-2018_subset.txt")
+f_sg <- file.path(ebird_dir, "ebd_relMay-2018_SG.txt")
+f_ru <- file.path(ebird_dir, "ebd_relMay-2018_rollup.txt")
 # sampling
-s_in <- file.path(ebird_dir, "ebd_sampling_relFeb-2018.txt")
-s_subset <- file.path(ebird_dir, "ebd_sampling_relFeb-2018_subset.txt")
-s_sg <- file.path(ebird_dir, "ebd_sampling_relFeb-2018_SG.txt")
+s_in <- file.path(ebird_dir, "ebd_sampling_relMay-2018.txt")
+s_subset <- file.path(ebird_dir, "ebd_sampling_relMay-2018_subset.txt")
+s_sg <- file.path(ebird_dir, "ebd_sampling_relMay-2018_SG.txt")
 
 # filter
 filters <- auk_ebd(f_in) %>%
-  auk_species(species = c("Gray Jay", "Blue Jay", "Steller's Jay", "Green Jay")) %>%
-  auk_country(country = c("US", "Canada", "Mexico", "Belize", "Guatemala", "Honduras", "Panama", "Costa Rica", "El Salvador")) %>%
+  auk_species(species = c("Gray Jay", "Blue Jay", 
+                          "Steller's Jay", "Green Jay")) %>%
+  auk_country(country = c("US", "Canada", "Mexico", "Belize", 
+                          "Guatemala", "Honduras", "Panama", 
+                          "Costa Rica", "El Salvador")) %>%
   auk_date(date = c("2010-01-01", "2012-12-31")) %>%
   auk_time(start_time = c("06:00", "12:00")) %>%
   auk_duration(duration = c(0, 120))
@@ -27,7 +30,7 @@ if (!file.exists(f_subset)) {
 
 x <- read_tsv(f_subset, quote = "", 
               col_types = cols(.default = col_character())) %>% 
-  select(-X49)
+  select(-X47)
 # evenly sample species
 set.seed(1)
 n_min <- min(table(x$`SCIENTIFIC NAME`))
@@ -38,10 +41,7 @@ x <- x %>%
 # sample to 500 records, make sure to get some from central america
 y1 <- sample_n(x %>% filter(!`COUNTRY CODE` %in% c("CA", "US")), 100)
 y2 <- sample_n(x %>% filter(`COUNTRY CODE` %in% c("CA", "US")), 400)
-y <- bind_rows(y1, y2) %>% 
-  # fill with fake names for privacy
-  mutate(`FIRST NAME` = stri_trans_general(`FIRST NAME`, "latin-ascii"),
-         `LAST NAME` = "eBirder")
+y <- bind_rows(y1, y2)
 f <- "inst/extdata/ebd-sample.txt"
 write_tsv(y, f, na = "")
 # remove any non-ascii characters
[…]
