### 2018-07-10 #136 @mstrimas
Sorry @stefaniebutland, I've been out of town off and on the last couple weeks, and won't be back until the end of next week, so this slipped through the cracks. If you feel what I already put together looks good for a blog post, I'd be happy to use that: https://github.com/mstrimas/auk-blog-post/blob/master/auk.md

What needs to be done to that to get it ready for posting?
