### 2018-07-12 #136 @stefaniebutland
That reads nicely as a blog post @mstrimas. Please read on when you have time.

I have a couple of suggestions for edits. When you're ready, please submit a pull request as outlined here: https://github.com/ropensci/roweb2#contributing-a-blog-post. Use date 2018-08-07 and if it is ready sooner and a spot opens up, we can change the date and publish sooner.

- would be helpful at start to note some of the types of info eBird contains - after this text I think "eBird is an online tool for recording bird observations. The eBird database currently contains over 500 million records of bird sightings, spanning every country and over 98% of species, making it an extremely valuable resource for bird research and conservation. " If not too much, it might entice readers if you give an example or two of the types of things birders would want to know from the data. Maybe refer to the examples you give at the end e.g. map species, so the reader knows there's a nice outcome at the end of the drier parts of the code explanations.

- In 2nd para of Accessing the data, the word "access" is used a lot. I know it's necessary but maybe can be reduced.

- Perhaps number the headings according to the numbers of 5 tasks under "Using auk".

- turn "Applications", "Presence-only data", "Zero-filled data" into more enticing headings that relate to your examples

- "One of the most obvious things to do with the presence data is make a map!" Please explain what your examples show with a line or two of text after the images. And be clear what you mean by effort in "The above map doesn't account for effort"

- I see you have a YouTube video about using `auk`. Would be good to find a spot to link to it.

@maelle is going to refer to using `auk` in an upcoming blog post so she will also review your draft, possibly once it's submitted as a pull request.

Hope this is helpful. Happy to answer any questions.
