+++ R/auk-distance.r
@@ -21,6 +21,7 @@
 #'
 #' @return An `auk_ebd` object.
 #' @export
+#' @family filter
 #' @examples
 #' # only keep checklists that are less than 10 km long
 #' system.file("extdata/ebd-sample.txt", package = "auk") %>%
