+++ R/auk-sampling.r
@@ -22,6 +22,7 @@
 #' @return An `auk_sampling` object storing the file reference and the desired
 #'   filters once created with other package functions.
 #' @export
+#' @family objects
 #' @examples
 #' # get the path to the example data included in the package
 #' # in practice, provide path to the sampling event data
