### 2018-07-19 ac01a2a
fixing normalizePath bug on Windows
