+++ README.md
@@ -154,7 +154,7 @@ drops these erroneous records and removes the blank column.
     f_out <- tempfile()
     # remove problem records
     auk_clean(f, f_out)
-    #> [1] "/var/folders/mg/qh40qmqd7376xn8qxd6hm5lwjyy0h2/T//RtmpJ2WgKZ/file7c6b265ea88d"
+    #> [1] "/var/folders/mg/qh40qmqd7376xn8qxd6hm5lwjyy0h2/T//RtmpCAWUPn/file8f41c2f9c7c"
     # number of lines in input
     length(readLines(f))
     #> [1] 51
