+++ ebd_filtered_grja.txt
@@ -1,0 +1,48 @@
+GLOBAL UNIQUE IDENTIFIER	LAST EDITED DATE	TAXONOMIC ORDER	CATEGORY	COMMON NAME	SCIENTIFIC NAME	SUBSPECIES COMMON NAME	SUBSPECIES SCIENTIFIC NAME	OBSERVATION COUNT	BREEDING BIRD ATLAS CODE	BREEDING BIRD ATLAS CATEGORY	AGE/SEX	COUNTRY	COUNTRY CODE	STATE	STATE CODE	COUNTY	COUNTY CODE	IBA CODE	BCR CODE	USFWS CODE	ATLAS BLOCK	LOCALITY	LOCALITY ID	LOCALITY TYPE	LATITUDE	LONGITUDE	OBSERVATION DATE	TIME OBSERVATIONS STARTED	OBSERVER ID	SAMPLING EVENT IDENTIFIER	PROTOCOL TYPE	PROTOCOL CODE	PROJECT CODE	DURATION MINUTES	EFFORT DISTANCE KM	EFFORT AREA HA	NUMBER OBSERVERS	ALL SPECIES REPORTED	GROUP IDENTIFIER	HAS MEDIA	APPROVED	REVIEWED	REASON	TRIP COMMENTS	SPECIES COMMENTS
+URN:CornellLabOfOrnithology:EBIRD:OBS120234325	2013-11-13 15:49:27	18722	species	Gray Jay	Perisoreus canadensis			3				Canada	CA	Manitoba	CA-MB	Churchill and Northern Manitoba	CA-MB-EE	CA-MB_003	7			Churchill--Twin Lakes Rd.	L349501	H	58.69846	-93.85119	2011-06-02	10:30:00	obsr269269	S8355489	Traveling	P22	EBIRD	60	8.047		4	1	G268135	0	1	0		Biking; warm, about 60-65F	
+URN:CornellLabOfOrnithology:EBIRD:OBS160958943	2017-10-19 09:59:32	18722	species	Gray Jay	Perisoreus canadensis			1				Canada	CA	Alberta	CA-AB	Calgary	CA-AB-SI		11			Horse Creek Road Marshes	L1594810	H	51.2826655	-114.5408776	2012-06-02	08:00:00	obsr131954	S11346052	Traveling	P22	EBIRD_CAN	120	5		8	1		0	1	0			
+URN:CornellLabOfOrnithology:EBIRD:OBS124851664	2017-08-15 16:14:25	18722	species	Gray Jay	Perisoreus canadensis			2				Canada	CA	Manitoba	CA-MB	Churchill and Northern Manitoba	CA-MB-EE		7			atlas square: 15VD67 - Owl River_Silcox Creek	L1260225	P	57.4708769	-93.6669112	2011-06-19	09:19:00	obsr121883	S8720531	Traveling	P22	EBIRD	5	12.874		1	1		0	1	0		Manitoba Breeding Bird Atlas Daily Observations	
+URN:CornellLabOfOrnithology:EBIRD:OBS587755907	2018-03-14 05:34:59	18722	species	Gray Jay	Perisoreus canadensis			2				Canada	CA	Quebec	CA-QC	Temiscouata	CA-QC-TT		14			Saint-Elzear-de-Temiscouata	L7063330	P	47.58333	-69.1	2010-09-21	06:30:00	obsr516205	S43636296	Historical	P62	EBIRD_QC	90			1	1		0	1	0		Feuillet d'EPOQ no 869453. Observateur : Rejean Deschenes.	
[…]
