+++ cran-comments.md
@@ -5,9 +5,9 @@ Fixes bug preventing state filtering from working and removes all non-ASCII char
 # Test environments
 
 - local OS X install, R 3.5.1
-- OS X (travis-ci), R 3.4.4
-- ubuntu 14.04 (travis-ci), R 3.4.4
-- Windows (appveyor), R 3.4.4
+- OS X (travis-ci), R 3.5.0
+- ubuntu 14.04 (travis-ci), R 3.5.0
+- Windows (appveyor), R 3.5.1
 - win-builder
 
 # R CMD check results
