### 2018-07-20 7076d5c
normalizePath issue
