+++ R/zzz.r
@@ -5,5 +5,5 @@
   #tax_date <- auk_version_date()["taxonomy_date"]
   #start_date <- format(as.Date(tax_date), "%b %Y")
   #end_date <- format(as.Date(tax_date + 365), "%b %Y")
-  packageStartupMessage(sprintf(m, "auk 0.2.0", "March 15, 2018"))
+  packageStartupMessage(sprintf(m, "auk 0.2.2", "March 15, 2018"))
 }
