### 2018-07-20 af32a0f
fixing issue with system2() on windows, needed to add a stderr file for it to work
