### 2018-07-20 07ec1e4
updating NEWS
