+++ NEWS.md
@@ -1,9 +1,10 @@
 # auk 0.2.2
 
 - Updated to work with EDB version 1.9
-- Fixed bug causing auk to fail on files downloaded via custom download form
 - Modified tests to be more general to all sample data
-- Fixed bug with normalizePath use on Windows
+- Fixed bug causing auk to fail on files downloaded via custom download form
+- Fixed bug with normalizePath() use on Windows
+- Fixed bug with system2() on Windows
 
 # auk 0.2.1
 
