### 2018-07-20 2e1b6b8
removing accidental output file
