+++ man/ebird_species.Rd
@@ -4,29 +4,29 @@
 \alias{ebird_species}
 \title{Lookup species in eBird taxonomy}
 \usage{
-ebird_species(x, scientific = TRUE)
+ebird_species(x, type = c("scientific", "common", "code"))
 }
 \arguments{
 \item{x}{character; species to look up, provided as scientific or
 English common names, or a mixture of both. Case insensitive.}
 
-\item{scientific}{logical; whether to return scientific (\code{TRUE}) or English
-common names (\code{FALSE}).}
+\item{type}{character; whether to return scientific names (\code{scientific}),
+English common names (\code{common}), or 6-letter eBird species codes (\code{code}).}
 }
 \value{
-Character vector of scientific names or common names if names if
-\code{scientific = FALSE}.
+Character vector of species identified by scientific name, common
+name, or species code.
 }
 \description{
 Given a list of common or scientific names, check that they appear in the
-official eBird taxonomy and convert them all to scientific names, or common
-names if \code{scientific = FALSE}. Un-matched species are returned as \code{NA}.
+official eBird taxonomy and convert them all to scientific names, common
+names, or species codes. Un-matched species are returned as \code{NA}.
 }
 \examples{
 # mix common and scientific names, case-insensitive
 species <- c("Blackburnian Warbler", "Poecile atricapillus",
              "american dipper", "Caribou")
-# species not in the ebird taxonomy return NA
+# note that species not in the ebird taxonomy return NA
 ebird_species(species)
 }
 \seealso{
