### 2018-07-23 1ede5d3
CRAN 0.2.2 release
