### 2018-07-31 1bd15df
BCR filter added
