+++ man/auk_extent.Rd
@@ -39,12 +39,13 @@ ebd <- auk_ebd(system.file("extdata/ebd-sample.txt", package = "auk"))
 auk_extent(ebd, extent = c(-125, 37, -120, 52))
 }
 \seealso{
-Other filter: \code{\link{auk_breeding}},
-  \code{\link{auk_complete}}, \code{\link{auk_country}},
-  \code{\link{auk_date}}, \code{\link{auk_distance}},
-  \code{\link{auk_duration}}, \code{\link{auk_filter}},
-  \code{\link{auk_last_edited}}, \code{\link{auk_project}},
-  \code{\link{auk_protocol}}, \code{\link{auk_species}},
-  \code{\link{auk_state}}, \code{\link{auk_time}}
+Other filter: \code{\link{auk_bcr}},
+  \code{\link{auk_breeding}}, \code{\link{auk_complete}},
+  \code{\link{auk_country}}, \code{\link{auk_date}},
+  \code{\link{auk_distance}}, \code{\link{auk_duration}},
+  \code{\link{auk_filter}}, \code{\link{auk_last_edited}},
+  \code{\link{auk_project}}, \code{\link{auk_protocol}},
+  \code{\link{auk_species}}, \code{\link{auk_state}},
+  \code{\link{auk_time}}
 }
 \concept{filter}
