+++ man/ebird_taxonomy.Rd
@@ -33,7 +33,8 @@ Non-ASCII characters (e.g. those with accents) have been converted to ASCII
 equivalents in this data frame.
 }
 \seealso{
-Other data: \code{\link{ebird_states}}
+Other data: \code{\link{bcr_codes}},
+  \code{\link{ebird_states}}
 }
 \concept{data}
 \keyword{datasets}

[…]
