### 2018-08-01 2617463
EBD_PATH now specifies data directory to increase code portability
