+++ R/auk-get-awk-path.r
@@ -1,15 +1,17 @@
-#' OS specific path to AWK
+#' OS specific path to AWK executable
 #'
-#' Return the OS specific path to AWK, or highlights if it's not installed. To 
-#' manually set the path to AWK, set the `AWK_PATH` environment variable in your
-#' `.Renviron` file.
+#' Return the OS specific path to AWK (e.g. `"C:/cygwin64/bin/gawk.exe"` or
+#' `"/usr/bin/awk"`), or highlights if it's not installed. To manually set the
+#' path to AWK, set the `AWK_PATH` environment variable in your `.Renviron`
+#' file, which can be accomplished with the helper function
+#' `auk_set_awk_path(path)`.
 #'
 #' @return Path to AWK or `NA` if AWK wasn't found.
 #' @export
-#' @family helpers
+#' @family paths
 #' @examples
-#' auk_getpath()
-auk_getpath <- function() {
+#' auk_get_awk_path()
+auk_get_awk_path <- function() {
   sysname <- tolower(Sys.info()[["sysname"]])
 
   # manually specified path
