+++ R/auk-ebd.r
@@ -3,14 +3,17 @@
 #' Create a reference to an eBird Basic Dataset (EBD) file in preparation for
 #' filtering using AWK.
 #'
-#' @param file character; input file.
+#' @param file character; input file. If file is not found as specified, it will
+#'   be looked for in the directory specified by the `EBD_PATH` environment
+#'   variable.
 #' @param file_sampling character; optional input sampling event data (i.e.
 #'   checklists) file, required if you intend to zero-fill the data to produce a
 #'   presence-absence data set. This file consists of just effort information
 #'   for every eBird checklist. Any species not appearing in the EBD for a given
 #'   checklist is implicitly considered to have a count of 0. This file should
 #'   be downloaded at the same time as the basic dataset to ensure they are in
-#'   sync.
+#'   sync. If file is not found as specified, it will be looked for in the
+#'   directory specified by the `EBD_PATH` environment variable.
 #' @param sep character; the input field separator, the eBird data are tab
 #'   separated so this should generally not be modified. Must only be a single
 #'   character and space delimited is not allowed since spaces appear in many of
@@ -50,10 +53,9 @@
 auk_ebd <- function(file, file_sampling, sep = "\t") {
   # checks
   assertthat::assert_that(
-    file.exists(file),
     assertthat::is.string(sep), nchar(sep) == 1, sep != " "
   )
-
+  file <- ebd_file(file)
   # read header rows
   header <- tolower(get_header(file, sep))
   header <- stringr::str_replace_all(header, "_", " ")
@@ -92,10 +94,7 @@ auk_ebd <- function(file, file_sampling, sep = "\t") {
 
   # process sampling data header
   if (!missing(file_sampling)) {
-    assertthat::assert_that(
-      file.exists(file_sampling)
-    )
-    file_sampling <- normalizePath(file_sampling)
+    file_sampling <- ebd_file(file_sampling)
     # variables not in sampling data
     not_in_sampling <- c("species", "breeding")
     filter_cols_sampling <- filter_cols[!filter_cols$id %in% not_in_sampling, ]
@@ -119,7 +118,7 @@ auk_ebd <- function(file, file_sampling, sep = "\t") {
   # output
   structure(
     list(
-      file = normalizePath(file),
+      file = file,
       file_sampling = file_sampling,
       output = NULL,
       output_sampling = NULL,
