+++ R/zzz.r
@@ -1,9 +1,11 @@
 .onAttach <- function(libname, pkgname) {
   m <- paste0("%s functions with EBD files downloaded after %s.\n",
               "See ?auk_version_date for details.")
-  #auk_version <- utils::paste("auk", packageVersion("auk"))
-  #tax_date <- auk_version_date()["taxonomy_date"]
-  #start_date <- format(as.Date(tax_date), "%b %Y")
-  #end_date <- format(as.Date(tax_date + 365), "%b %Y")
-  packageStartupMessage(sprintf(m, "auk 0.2.2", "March 15, 2018"))
+  if (is.na(auk_get_ebd_path())) {
+    m <- paste(m, "\nNo EBD data directory set, see ?auk_set_ebd_path to set",
+               "EBD_PATH")
+  } else {
+    m <- paste(m, "\nEBD data directory: ", auk_get_ebd_path())
+  }
+  packageStartupMessage(sprintf(m, "auk 0.2.3", "March 15, 2018"))
 }
