### 2018-08-03 d1628bb
checking if Cygwin works on Appveyor🤞
