### 2018-08-03 b296b3f
remove skip_on_windows() from tests; clarify that extent = bounding box
