+++ R/auk-protocol.r
@@ -7,9 +7,10 @@
 #'
 #' @param x `auk_ebd` or `auk_sampling` object; reference to file created by 
 #'   [auk_ebd()] or [auk_sampling()].
-#' @param protocol character; "Stationary", "Traveling", or "Incidental". Other 
-#'   protocols exist in the database, however, this function only extracts these 
-#'   three standard protocols. Multiple protocols are accepted.
+#' @param protocol character; "Stationary", "Traveling", "Area", or
+#'   "Incidental". Other protocols exist in the database, however, this function
+#'   only extracts these three standard protocols. Multiple protocols are
+#'   accepted.
 #' 
 #' @details This function can also work with on an `auk_sampling` object if the 
 #'   user only wishes to filter the sampling event data.
@@ -32,7 +33,7 @@ auk_protocol <- function(x, protocol)  {
 #' @export
 auk_protocol.auk_ebd <- function(x, protocol) {
   assertthat::assert_that(
-    all(protocol %in% c("Stationary", "Traveling", "Incidental"))
+    all(protocol %in% c("Stationary", "Traveling", "Area", "Incidental"))
   )
   
   # set filter list
