### 2018-08-08 75c6dfd
path setting functions now change .Renviron rather than opening it
