### 2018-08-09 b2f3698
auk_extent() renamed auk_bbox(), closes issue #21, thanks @maelle!
