### 2018-08-13 75d4dbe
apply EBD_PATH to auk_sampling()
