### 2018-08-21 740f017
zerofill now throws an error if incomplete checklists used
