+++ tests/testthat/test_auk-zerofill.r
@@ -87,10 +87,9 @@ test_that("auk_zerofill error is auk_unique() hasn't been run", {
 test_that("auk_zerofill lack of complete checklists throws error", {
   f_ebd <- system.file("extdata/zerofill-ex_ebd.txt", package = "auk")
   f_smpl <- system.file("extdata/zerofill-ex_sampling.txt", package = "auk")
-  ebd <- auk_ebd(f_ebd, f_smpl)
-  ebd$output <- f_ebd
-  ebd$output_sampling <- f_smpl
-
+  ebd <- read_ebd(f_ebd)
+  smpl <- read_sampling(f_smpl)
+  smpl$all_species_reported[sample(1:nrow(smpl), 3)] <- FALSE
   expect_error(auk_zerofill(ebd))
 })
 
