+++ inst/extdata/zerofill-ex_ebd.txt
@@ -2,7 +2,6 @@ GLOBAL UNIQUE IDENTIFIER	LAST EDITED DATE	TAXONOMIC ORDER	CATEGORY	COMMON NAME	S
 URN:CornellLabOfOrnithology:EBIRD:OBS230331527	2017-05-27 10:25:40	9841	species	Collared Kingfisher	Todiramphus chloris			1				Singapore	SG	Singapore	SG-							Seletar Dam	L1403154	H	1.423116	103.8629908	2012-10-02	17:10:00	obsr303417	S16687187	Stationary	P21	EBIRD_AU	35			2	1		0	1	0		With Con Foley, on the dam wall, looking for an Oriental Plover that had been seen the day before (rarity for Singapore) and Malaysian Plover (one pair is reputedly resident). Eremaea Birds List 145689	
 URN:CornellLabOfOrnithology:EBIRD:OBS137367921	2017-05-27 09:21:26	9841	species	Collared Kingfisher	Todiramphus chloris			2				Singapore	SG	Singapore	SG-							Lower Peirce Reservoir Park	L948606	H	1.3706712	103.8262081	2012-01-24	08:00:00	obsr204697	S9681625	Traveling	P22	EBIRD	120	1.609		1	1		0	1	0			
 URN:CornellLabOfOrnithology:EBIRD:OBS164948143	2017-05-27 11:00:13	9841	species	Collared Kingfisher	Todiramphus chloris			1				Singapore	SG	Singapore	SG-							Changi Village Hotel	L1277940	P	1.3905938	103.9854455	2012-09-27	17:30:00	obsr189448	S11675087	Traveling	P22	EBIRD	40	.805		1	1		0	1	0			
-URN:CornellLabOfOrnithology:EBIRD:OBS574442604	2018-02-21 02:14:09	9724	species	White-throated Kingfisher	Halcyon smyrnensis			3				Singapore	SG	Singapore	SG-							Seletar Dam	L1403154	H	1.423116	103.8629908	2012-09-02	07:30:00	obsr775526	S42504549	Historical	P62	EBIRD	180			6	0		0	1	0		with Doreen, Ann Ang, John Ang, Jacky Soh and Peng Yuan Han	
 URN:CornellLabOfOrnithology:EBIRD:OBS291206773	2017-05-27 10:36:20	9841	species	Collared Kingfisher	Todiramphus chloris			X				Singapore	SG	Singapore	SG-							NUS campus, Kent Ridge	L3294638	P	1.2990971	103.7750959	2012-02-03	08:00:00	obsr330987	S21361601	Traveling	P22	EBIRD	120	1		1	1		0	1	0			
 URN:CornellLabOfOrnithology:EBIRD:OBS230143921	2017-05-27 10:25:40	9841	species	Collared Kingfisher	Todiramphus chloris			2				Singapore	SG	Singapore	SG-							Botanic Gardens	L2534689	P	1.31	103.81633	2012-05-28		obsr19060	S16575332	Incidental	P20	EBIRD_AU					1		0	1	0		Also unable to separate Swiftlets between Black-nest and Edible-nest.  Mute swans are believed to be captive. Eremaea Birds List 132429	At least 2 and probably more.
 URN:CornellLabOfOrnithology:EBIRD:OBS409866559	2017-05-27 10:25:40	9724	species	White-throated Kingfisher	Halcyon smyrnensis			1				Singapore	SG	Singapore	SG-							Lower Peirce Reservoir Park	L948606	H	1.3706712	103.8262081	2012-08-08	17:00:00	obsr495297	S30081693	Traveling	P22	EBIRD	90	1.931		2	1		0	1	0			
[…]
